+++ src/extensions/export.ts
@@ -11,7 +11,7 @@ import { FirebaseError } from "../error";
 import { promptOnce } from "../prompt";
 import { parseSecretVersionResourceName, toSecretVersionResourceName } from "../gcp/secretManager";
 import { getActiveSecrets } from "./secretsUtils";
-
+import { writeEnvFiles, writeExtensionsToFirebaseJson } from "./manifestHelper";
 /**
  * parameterizeProject searchs spec.params for any param that include projectId or projectNumber,
  * and replaces it with a parameterized version that can be used on other projects.
@@ -80,56 +80,5 @@ function displaySpecs(specs: InstanceSpec[]): void {
       logger.info(`\t${p[0]}=${p[1]}`);
     }
     logger.info("");
-  }
-}
-
-function writeExtensionsToFirebaseJson(have: InstanceSpec[], existingConfig: Config): void {
-  const extensions = existingConfig.get("extensions", {});
-  for (const s of have) {
-    extensions[s.instanceId] = refs.toExtensionVersionRef(s.ref!);
-  }
-  existingConfig.set("extensions", extensions);
-  logger.info("Adding Extensions to " + clc.bold("firebase.json") + "...");
-  existingConfig.writeProjectFile("firebase.json", existingConfig.src);
-}
-
-async function writeEnvFile(spec: InstanceSpec, existingConfig: Config, force?: boolean) {
-  const content = Object.entries(spec.params)
-    .map((r) => `${r[0]}=${r[1]}`)
-    .join("\n");
-  await existingConfig.askWriteProjectFile(`extensions/${spec.instanceId}.env`, content, force);
-}
-
-export async function writeFiles(have: InstanceSpec[], options: Options) {
-  const existingConfig = Config.load(options, true);
-  if (!existingConfig) {
-    throw new FirebaseError(
-      "Not currently in a Firebase directory. Please run `firebase init` to create a Firebase directory."
-    );
-  }
-  if (
-    existingConfig.has("extensions") &&
-    Object.keys(existingConfig.get("extensions")).length &&
-    !options.nonInteractive &&
-    !options.force
-  ) {
-    const currentExtensions = Object.entries(existingConfig.get("extensions"))
-      .map((i) => `${i[0]}: ${i[1]}`)
-      .join("\n\t");
-    const overwrite = await promptOnce({
-      type: "list",
-      message: `firebase.json already contains extensions:\n${currentExtensions}\nWould you like to overwrite or merge?`,
-      choices: [
-        { name: "Overwrite", value: true },
-        { name: "Merge", value: false },
-      ],
-    });
-    if (overwrite) {
-      existingConfig.set("extensions", {});
-    }
-  }
-  writeExtensionsToFirebaseJson(have, existingConfig);
-  for (const spec of have) {
-    await writeEnvFile(spec, existingConfig, options.force);
   }
 }
